Changed how pieces are initialized
The new CustomTypes script is preloaded by Accessor as a constant, and can be referenced by pieces when resolving their starting states. This system allows for pieces to reference other nodes in their starting states while still having const state_forms.

Teams are now optional for every PieceType except King2i. When a team is not present, the PieceType.can_take() function will default to true. Teams can be manually added to a Piece's starting state as a NodePath, which will automatically be converted into the Team Node.