Added primitive check system
Options will be erased if another option on the next turn results in _evaluate() returning true.

Checkmate would be a loss of all options in Chess

Check would be a loss of the trivial options in Chess. Not sure how to implement this, since options are pretty explicitly property of Pieces, and making a "trivial piece" sounds weird.